#10 stat choice will now display all the tournaments existing in the data
